Ran python-modernize

+++ rdiff-backup/rdiff_backup/Rdiff.py
@@ -19,8 +19,9 @@
 
 """Invoke rdiff utility to make signatures, deltas, or patch"""
 
+from __future__ import absolute_import
 import os, librsync
-import Globals, log, static, TempFile, rpath, hash
+from . import Globals, log, static, TempFile, rpath, hash
 
 
 def get_signature(rp, blocksize = None):
@@ -40,7 +41,7 @@ def find_blocksize(file_len):
 	"""
 	if file_len < 4096: return 64 # set minimum of 64 bytes
 	else: # Use square root, rounding to nearest 16
-		return long(pow(file_len, 0.5)/16)*16
+		return int(pow(file_len, 0.5)/16)*16
 
 def get_delta_sigfileobj(sig_fileobj, rp_new):
 	"""Like get_delta but signature is in a file object"""
